bin/utils/requirements.txt
ipython[nbconvert]>=3.1
biopython

